Applied the review suggestion
The most important is a change to SecondaryIndexCheckpointTracker, thanks to help from @timothycoleman, it simplifies the threading coordination. Now the checkpoint tracking is started automatically when tracker is created. This is fine, as if there are no changes in batch, then no commit is made.

This change requires more testing, but this will come in the follow up.