[KDB-872] Add logging for significant garbage collections (#5134) (#5161)
* Add logging for significant garbage collections

To make it clear from the logs if slow messages or leader failovers are attributable to GC.

Execution engine suspensions longer than 48ms are logged as Information
Execution engine suspensions longer than 600ms are logged as Warnings
Full compacting GC start/end are logged as Information.

These will be logged even if the node shortly goes offline for truncation, preventing the EE suspension from appearing in the metrics.

If GC is determined as the cause, a likely sensible course of action could be to reduce the Stream Info Cache Capacity (say, to the 100k traditional value) and/or consider enabling ServerGC.