[KDB-916] Add TCP Plugin Configuration Option for TCP read expiry: TcpReadTimeoutMs (#5144) (#5167)
* Add TCP Plugin Configuration Option for TCP read expiry: TcpReadTimeoutMs (#5144)

In gRPC this is already driven by the request itself. In TCP we have no such mechanism. We opt for this solution instead of adding it to the TCP protocol to make the change smaller, safer, and easier to roll out while still being useful.

This is a forward port of #5144. It's a little bit different because in this version TCP is a plugin in a separate repository.

* Include the lifetime of the message when logging expiries