Fix authenticated_requests_made_from_a_follower flaky test
The test writes to a follower node which forwards the write to the leader.

The reason the test was flaky is that it waits for the admin user to be created on the follower node, but this doesn't guarantee that the user is indexed on the leader.

The test fails when the request gets forwarded to the leader but the user isn't yet indexed, at which point the request is rejected.

The fix is to wait for the admin user to be created on the leader as part of spinning up the cluster.