[DB-1655] Add forwards compatibility with new persistent subscription checkpoints (#5244)
* Add forwards compatibility with new persistent subscription checkpoints

24.10 writes Persistent Subscription Checkpoints prefixed with a $.
23.10.7 will not find these.

This PR makes it so that 23.10.8 will find them.

Prior to this PR, if 23.10 opened a database from 24.10, the checkpoint would not be found and the persistent subscription would start from the beginning, which could be slow.

(cherry picked from commit 6e21abb54be4ead53beaeccee23d82a759ebd1ae)

* Fix build following docs move