Adjust options processing for net10 breaking change
Before net10 nulls were treated as having no configured value.
Since net10 nulls are respected as configuration values in their own right.

https://learn.microsoft.com/en-us/dotnet/core/compatibility/extensions/10.0/configuration-null-values-preserved

The problem for us was that if the GossipSeed is configured as an array

```
GossipSeed: [ "node1.eventstore:2111", "node2.eventstore:2112" ]
```

then the configuration became

```
key           value
GossipSeed    null
GossipSeed:0  "host1:2113"
GossipSeed:1  "host2:2113"
```

but the null GossipSeed value is now respected and so the 0,1 subsections are not checked, leaving the gossip seed value as null.