Fix projection stall race condition when writing empty transactions
If a projection had finished processing all events in its input stream, and then another event was written to its input stream, under certain conditions a race condition meant that the projection would remain idle and not process that new event.

The projection would process that new event (and any subsequent events) if any event was written to any stream. i.e. the stalled projection could be recovered by writing any event to any stream.

The stall requires transactions without any events in to be written.

When a projection reaches the end of its input stream, it subscribes to the AwakerService to be notified when there are more events to process.

When a new event is committed, AwakerService decides which subscribers need to be notified. However, it only notifies them when the indexing process reaches the end of the log or a threshold is reached. (This prevents unnecessary notifications and resubscriptions).

Unfortunately, the AwakerService was only notified when the indexing process reaches the end of the log if the last record in the log was an event. This was almost always the case, but is not the case if the last record is an empty transaction containing no events (i.e. the client sent an append request with 0 events in).

A stall could therefore occur if a projection has subscribed to the AwakerService, then an event E that it is interested in is written, immediately followed by an empty transaction before E is indexed (this is the race). The AwakerService would then know that the projection needs notifying, but will not notify it, waiting instead to be notified that the indexing process has reached the end of the log, which it never will, unless another event is written.

This PR fixes the race condition by having the indexing process send a notification when the end of the log is reached, even if the last record is not an event.